Crate pagespeedonline2

This documentation was generated from pagespeedonline crate version 0.0.1. The original source code can be found on github.

Features

Handle the following Resources with ease from the central hub ...

pagespeedapi (runpagespeed)

Everything else about the pagespeedonline v2 API can be found at the official documentation site.

Not what you are looking for ? Find all other google APIs in their Rust documentation index.

Structure of this Library

The API is structured into the following primary items:

Hub

a central object to maintain state and allow accessing all Activities

Resources

primary types that you can apply Activities to

a collection of properties and Parts

Parts

a collection of properties

never directly used in Activities

Activities

operations to apply to Resources

Generally speaking, you can invoke Activities like this:

let r = hub.resource().activity(...).doit()

Or specifically ...

let r = hub.pagespeedapi().runpagespeed(...).doit()

The resource() and activity(...) calls create builders. The second one dealing with Activities supports various methods to configure the impending operation (not shown here). It is made such that all required arguments have to be specified right away (i.e. (...)), whereas all optional ones can be build up as desired. The doit() method performs the actual communication with the server and returns the respective result.

Usage

Setting up your Project

To use this library, you would put the following lines into your Cargo.toml file:

[dependencies]
pagespeedonline2 = "0.0.1"

A complete example

extern crate hyper;
extern crate "yup-oauth2" as oauth2;
extern crate "rustc-serialize" as rustc_serialize;
extern crate pagespeedonline2;
use pagespeedonline2::cmn::Result;
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use pagespeedonline2::Pagespeedonline;
 
// Get an ApplicationSecret instance by some means. It contains the `client_id` and 
// `client_secret`, among other things.
let secret: ApplicationSecret = Default::default();
// Instantiate the authenticator. It will choose a suitable authentication flow for you, 
// unless you replace  `None` with the desired Flow.
// Provide your own `AuthenticatorDelegate` to adjust the way it operates and get feedback about 
// what's going on. You probably want to bring in your own `TokenStorage` to persist tokens and
// retrieve them from storage.
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = Pagespeedonline::new(hyper::Client::new(), auth);
// You can configure optional parameters by calling the respective setters at will, and
// execute the final call using `doit()`.
// Values shown here are possibly random and not representative !
let result = hub.pagespeedapi().runpagespeed("url")
             .strategy("accusam")
             .screenshot(true)
             .add_rule("justo")
             .locale("amet.")
             .filter_third_party_resources(false)
             .doit();
 
match result {
    Result::HttpError(err) => println!("HTTPERROR: {:?}", err),
    Result::MissingAPIKey => println!("Missing API Key"),
    Result::MissingToken => println!("Missing Token"),
    Result::Failure(_) => println!("General Failure (Response doesn't print)"),
    Result::FieldClash(clashed_field) => println!("FIELD CLASH: {:?}", clashed_field),
    Result::Success(_) => println!("Success (value doesn't print)"),
}

Handling Errors

All errors produced by the system are provided either as Result enumeration as return value of the doit() methods, or handed as possibly intermediate results to either the Hub Delegate, or the Authenticator Delegate.

When delegates handle errors or intermediate values, they may have a chance to instruct the system to retry. This makes the system potentially resilient to all kinds of errors.

About Customization/Callbacks

You may alter the way an doit() method is called by providing a delegate to the Method Builder before making the final doit() call. Respective methods will be called to provide progress information, as well as determine whether the system should retry on failure.

The delegate trait is default-implemented, allowing you to customize it with minimal effort.

About Parts

All structures provided by this library are made to be enocodable and decodable via json. Optionals are used to indicate that partial requests are responses are valid. Most optionals are are considered Parts which are identifyable by name, which will be sent to the server to indicate either the set parts of the request or the desired parts in the response.

About Builder Arguments

Using method builders, you are able to prepare an action call by repeatedly calling it's methods. These will always take a single argument, for which the following statements are true.

PODs are handed by copy

strings are passed as &str

request values are borrowed

Arguments will always be copied or cloned into the builder, to make them independent of their original life times.

Modules

cmn

Macros

map!
This macro is advertised in the documentation, which is why we deliver it as well

Structs

PagespeedApiFormatStringV2
There is no detailed description.

This type is not used in any activity, and only used as part of another schema.

PagespeedApiFormatStringV2Args
List of arguments for the format string.

This type is not used in any activity, and only used as part of another schema.

PagespeedApiFormatStringV2ArgsRects
The screen rectangles being referred to, with dimensions measured in CSS pixels. This is only ever used for SNAPSHOT_RECT arguments. If this is absent for a SNAPSHOT_RECT argument, it means that that argument refers to the entire snapshot.

This type is not used in any activity, and only used as part of another schema.

PagespeedApiFormatStringV2ArgsSecondaryRects
Secondary screen rectangles being referred to, with dimensions measured in CSS pixels. This is only ever used for SNAPSHOT_RECT arguments.

This type is not used in any activity, and only used as part of another schema.

PagespeedApiImageV2
There is no detailed description.

This type is not used in any activity, and only used as part of another schema.

PagespeedApiImageV2PageRect
The region of the page that is captured by this image, with dimensions measured in CSS pixels.

This type is not used in any activity, and only used as part of another schema.

PagespeedapiMethodsBuilder
A builder providing access to all methods supported on pagespeedapi resources. It is not used directly, but through the Pagespeedonline hub.

PagespeedapiRunpagespeedMethodBuilder
Runs PageSpeed analysis on the page at the specified URL, and returns PageSpeed scores, a list of suggestions to make that page faster, and other information.

Pagespeedonline
Central instance to access all Pagespeedonline related resource activities

Result
There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

runpagespeed pagespeedapi (response)

ResultFormattedResults
Localized PageSpeed results. Contains a ruleResults entry for each PageSpeed rule instantiated and run by the server.

This type is not used in any activity, and only used as part of another schema.

ResultFormattedResultsRuleResults
The enum-like identifier for this rule. For instance "EnableKeepAlive" or "AvoidCssImport". Not localized.

This type is not used in any activity, and only used as part of another schema.

ResultFormattedResultsRuleResultsUrlBlocks
List of blocks of URLs. Each block may contain a heading and a list of URLs. Each URL may optionally include additional details.

This type is not used in any activity, and only used as part of another schema.

ResultFormattedResultsRuleResultsUrlBlocksUrls
List of entries that provide information about URLs in the url block. Optional.

This type is not used in any activity, and only used as part of another schema.

ResultPageStats
Summary statistics for the page, such as number of JavaScript bytes, number of HTML bytes, etc.

This type is not used in any activity, and only used as part of another schema.

ResultRuleGroups
The name of this rule group: one of "SPEED" or "USABILITY".

This type is not used in any activity, and only used as part of another schema.

ResultVersion
The version of PageSpeed used to generate these results.

This type is not used in any activity, and only used as part of another schema.